// src/lib/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5"
  }
  public: {
    Tables: {
// ...
      clients: {
        Row: {
          address: string
          bin_number: string | null
          client_type: Database["public"]["Enums"]["client_type"] | null
          company_name: string
          contact_person: string
          created_at: string | null
          email: string
          id: string
          notes: string | null
          phone: string
          status: Database["public"]["Enums"]["client_status"] | null
          tin_number: string | null
          trade_license_number: string | null
          updated_at: string | null
        }
        Insert: {
          address: string
          bin_number?: string | null
          client_type?: Database["public"]["Enums"]["client_type"] | null
          company_name: string
          contact_person: string
          created_at?: string | null
          email: string
          id?: string
          notes?: string | null
          phone: string
          status?: Database["public"]["Enums"]["client_status"] | null
          tin_number?: string | null
          trade_license_number?: string | null
          updated_at?: string | null
        }
        Update: {
          address?: string
          bin_number?: string | null
          client_type?: Database["public"]["Enums"]["client_type"] | null
          company_name?: string
          contact_person?: string
          created_at?: string | null
          email?: string
          id?: string
          notes?: string | null
          phone?: string
          status?: Database["public"]["Enums"]["client_status"] | null
          tin_number?: string | null
          trade_license_number?: string | null
          updated_at?: string | null
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never